requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pytesseract==0.3.12
requests==2.32.3
pillow==11.1.0
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
